user auth added

+++ controllers/items.js
@@ -1,16 +1,23 @@
 const itemsRouter = require("express").Router();
 const Item = require("../models/item");
 const User = require("../models/user");
+const jwt = require("jsonwebtoken");
+const middleware = require("../utils/middleware");
 
 itemsRouter.get("/", async (req, res) => {
   const items = await Item.find({}).populate("user", { items: 0 });
   res.json(items);
 });
 
-itemsRouter.post("/", async (req, res) => {
+itemsRouter.post("/", middleware.getToken, async (req, res) => {
   const body = req.body;
 
-  const user = await User.findById(body.userId);
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
 
   const item = new Item({
     title: body.title,
@@ -40,9 +47,20 @@ itemsRouter.get("/:id", async (req, res) => {
   }
 });
 
-itemsRouter.delete("/:id", async (req, res) => {
-  await Item.findByIdAndRemove(req.params.id);
-  res.status(204).end();
+itemsRouter.delete("/:id", middleware.getToken, async (req, res) => {
+  const decodedToken = jwt.verify(req.token, process.env.SECRET);
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const item = await Item.findById(req.params.id);
+
+  if (item.user.toString() === decodedToken.id) {
+    await Item.findByIdAndRemove(req.params.id);
+    res.status(204).end();
+  } else {
+    res.status(401).json({ error: "user not allowed to delete this item" });
+  }
 });
 
 module.exports = itemsRouter;
